+++ chat/src/test/java/com/back/chat/OutboxRelayDeadLetterTest.java
@@ -1,6 +1,6 @@
 package com.back.chat;
 
-import com.back.chat.adapter.out.ChatOutboxRelay;
+import com.back.chat.adapter.out.outbox.ChatOutboxRelay;
 import com.back.chat.adapter.out.outbox.ChatOutbox;
 import com.back.chat.adapter.out.outbox.ChatOutboxRepository;
 import com.back.chat.adapter.out.outbox.OutboxStatus;
