+++ common/src/main/java/com/back/common/chat/ChatDeadLetterPayload.java
@@ -1,0 +1,15 @@
+package com.back.common.chat;
+
+import java.time.LocalDateTime;
+
+public record ChatDeadLetterPayload(
+        String source,
+        String eventType,
+        String eventId,
+        Long outboxId,
+        int retryCount,
+        String lastError,
+        LocalDateTime deadAt,
+        String originalPayload
+) {
+}
