+++ common/src/main/java/com/back/common/chat/ChatMessageBlindedKafkaEvent.java
@@ -1,6 +1,5 @@
 package com.back.common.chat;
 
-import com.back.common.event.EventName;
 
 import java.time.LocalDateTime;
 
@@ -8,5 +7,5 @@ public record ChatMessageBlindedKafkaEvent(
         String eventId,
         Long authorUserId,
         LocalDateTime createdAt
-) implements EventName {
+) {
 }
